Extracts trans tables into a separate library
