+++ library/src/trans_table/TransTableL.cpp
