+++ library/src/trans_table/TransTableS.cpp
